Add minimal chord diagram generator.
This is a Python script which expects to find files named 'chords.yaml'
and 'chord_base.svg' in the current directory.

For each chord described in the YAML file, using the SVG file as a
template, it creates a new SVG file for the new chord and saves the
resulting SVG as a new file.

These SVGs can then be rendered directly or exported to PNG using e.g.
headless Inkscape.