Form Inkscape command differently depending on version.
Inkscape v1.0.0 broke compatibility with the command line arguments in
the 0.9.x series. The '-z/--without-gui' option was deprecated, the '-e'
option was renamed to '-o', and we now have to specify an export type.

Rather than force the app to use either 0.9.x or >= 1.0.0, I decided to
execute `inkscape -V` and parse the version string, and then form up the
arguments to generate the diagram differently depending on what version
we have.