Add support for frets beyond the 4th.
As long as the full chord diagram can still be shown within a span of 4
frets, generate_chord_diagram will 'shift' along the fretboard as needed
to show the finger positions. When this happens, the fret labels are
changed to show the new fret numbers, and the first fret label is shown
in a more contrasted fill to draw attention to the change in fret
number.

So for example, the C minor chord (using the A minor shape at the 3rd
fret) uses frets 3, 4 and 5. So to show this, the diagram displays frets
3, 4, 5 and 6 instead of the usual 1, 2, 3 and 4, and the label for fret
3 is shown in a darker shade.

The web app still needs to be updated to allow this kind of diagram.