Fix bug with played notes shown on barres.
Previously, if a finger was barring, we would skip the entire code block
where we would normally place a dot marker, however, that is also the
code block where we set the played note for the string.

Move that code into a separate function, and call it before we check for
a barre and potentially skip over that iteration in the loop.